# doc/api.yaml
openapi: 3.1.0
info:
  title: WASAPhoto
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto!

    You can upload your photos directly from your PC, and they will be visible to everyone following you.
  version: 1.0.0
components:
  schemas:
    User:
      title: User
      description: This object represents a single user
      type: object
      properties:
        userId:
          type: integer
          description: Unique identifier for the user
          example: 1234
          readOnly: true
          minimum: 0
        username:
          description: The username displayed to the others
          type: string
          example: CoolUsername42
          minLength: 3
          maxLength: 32
          pattern: '^[A-Za-z0-9_\-]{3,32}$'
      required: [ "userId", "username" ]
    Profile:
      title: Profile
      description: This object represents the profile data of a user
      type: object
      properties:
        userId: { $ref: '#/components/schemas/User/properties/userId' }
        username: { $ref: '#/components/schemas/User/properties/username' }
        photoCount:
          description: The count of how many photos this user uploaded
          type: integer
          example: 42
          minimum: 0
        followersCount:
          description: The amount of users following this user
          type: integer
          example: 42
          minimum: 0
        followingCount:
          description: The amount of users followed by this user
          type: integer
          example: 42
          minimum: 0
        bannedCount:
          description: The amount of users banned by this user
          type: integer
          example: 42
          minimum: 0
      required: [ "userId", "username", "photoCount", "followersCount", "followingCount", "bannedCount" ]
...
    Error:
      title: Error
      description: |-
        This object represents a single error to be returned for certain requests
      type: object
      properties:
        code:
          type: integer
          description: An HTTP error code
          example: 404
        message:
          type: string
          description: A brief message describing the error code
          example: The specified resource was not found
      required: [ "code", "message" ]
  parameters:
    usernameParam:
      name: username
      in: path
      description: Username of the user
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/User/properties/username' }
    targetUsernameParam:
      name: targetUsername
      in: path
      description: Username of the target
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/User/properties/username' }
...
  responses:
...
    NoContent:
      description: The request was successful, but there is no content to return.
    # 400
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    # 401
    Unauthorized:
      description: Invalid or expired authentication credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    # 404
    NotFound:
      description: The specified resource was not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    # 409
...
    UnprocessableEntity:
      description: The request was well-formed but was unable to be followed due to semantic errors.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    # 500
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: [ ]
...
paths:
...
  /users/{username}/profile:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
    get:
      tags: [ "User" ]
      operationId: getUserProfile
      summary: Retrieve the information of a user
      description: |-
        Given a user's username, retrieve all the public info available.
      responses:
        "200":
          description: Successfully retrieved profile data
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "404": { $ref: '#/components/responses/NotFound' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
        "500": { $ref: '#/components/responses/InternalServerError' }
...
  /users/{username}/following/count:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
    get:
      tags: [ "Follow" ]
      operationId: getFollowingCount
      summary: Get the count of users followed by the current user
      description: Get the count of users followed by the current user
      responses:
        "200":
          description: Successfully retrieved the count of followed users
          content:
            application/json:
              schema:
                type: integer
                description: count of followed users
                example: 42
                minimum: 0
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /users/{username}/following/list:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
    get:
      tags: [ "Follow" ]
      operationId: getFollowingList
      summary: Get the list of users followed by the current user
      description: Get the list of users followed by the current user
      responses:
        "200":
          description: Successfully retrieved the list of followed users
          content:
            application/json:
              schema:
                type: array
                description: list of usernames
                items: { $ref: '#/components/schemas/User/properties/username' }
                uniqueItems: true
                minItems: 0
                maxItems: 99999
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /users/{username}/followers:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
    post:
      tags: [ "Follow" ]
      operationId: followUser
      summary: Follow a user
      description: Given a user's username, follow them.
      requestBody:
        description: The username of the user to follow
        required: true
        content:
          application/json:
            schema:
              type: object
              description: username schema
              properties:
                username: { $ref: '#/components/schemas/User/properties/username' }
      responses:
        "201":
          description: User followed successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /users/{username}/followers/count:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
    get:
      tags: [ "Follow" ]
      operationId: getFollowersCount
      summary: Get the count of users following the current user
      description: Get the count of users following the current user
      responses:
        "200":
          description: Successfully retrieved the count of followers
          content:
            application/json:
              schema:
                type: integer
                description: count of followers
                example: 42
                minimum: 0
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /users/{username}/followers/list:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
    get:
      tags: [ "Follow" ]
      operationId: getFollowersList
      summary: Get the list of users following the current user
      description: Get the list of users following the current user
      responses:
        "200":
          description: Successfully retrieved the list of followers
          content:
            application/json:
              schema:
                type: array
                description: list of usernames
                items: { $ref: '#/components/schemas/User/properties/username' }
                uniqueItems: true
                minItems: 0
                maxItems: 99999
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /users/{username}/followers/list/{targetUsername}:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
      - $ref: '#/components/parameters/targetUsernameParam'
    get:
      tags: [ "Follow" ]
      operationId: getFollowStatus
      summary: Check if a user is following another
      description: |-
        Given a user's username, check if they are following another user.
        In this case, the target user is the 1st parameter.
      responses:
        "200":
          description: User is following the target user
          content:
            application/json:
              schema:
                type: object
                properties:
                  isFollowing:
                    type: boolean
                    description: Whether the user is following the target user
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /users/{username}/followers/{targetUsername}:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
      - $ref: '#/components/parameters/targetUsernameParam'
    delete:
      tags: [ "Follow" ]
      operationId: unfollowUser
      summary: Unfollow a user
      description: Given a user's username, unfollow them.
      responses:
        "204": { $ref: '#/components/responses/NoContent' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
        "500": { $ref: '#/components/responses/InternalServerError' }
...